Raise warnings as errors
Well-behaved third party packages (and first party code) may raise
warnings when used in unexpected ways.

We have historically ignored these warnings, but we shouldn't. This
patch will convert warnings to errors during pytest runs, so that we are
alerted to potentially risky behaviour and have to consciously address
it.