BAU: use the latest pscyopg3 postgres adapter
Building psycopg2 on Mac OS throws errors around missing SSL libraries -
this is definitely solvable but while looking into it it looks like
psycopg3 is stable and OK to use. It builds without any further
workarounds and looks to be recommended:
> If you are developing something new, Psycopg 3 is the most current
implementation of the adapter.

SQL alchemy from version 2.0 onwards supports pscyopg3 but it looks like
you have to explicitly tell in when connecting that it should use the
latest version otherwise it defaults to trying to import psycopg2.

Reference
https://docs.sqlalchemy.org/en/20/dialects/postgresql.html#module-sqlalchemy.dialects.postgresql.psycopg