Add framework for populating flask config
Use `pydantic-settings` to pull our app configuration from the
environment.

For local development this _could_ be extended to pull from eg `.env`
files, but given we're planning to run this through docker-compose, we
should be able to rely on setting these values in `docker-compose.yml`
and having them injected as true environment variables.

By using `pydantic-settings`, we get a few nice benefits:

- Environment variables will be coerced based on the type hints on the
  class.
- If any declared fields are missing (ie environment variables are not
  present), the app will refuse to start. This should prevent us
  launching an app that has [totally] invalid or missing configuration.

Unfortunately these type hints do not filter all the way through the
Flask framework, so when accessing eg `app.config` or
`current_app.config`, we won't get intellisense/completion based on the
type hints and available attributes. This matches historical funding
service behaviour, so it's not _worse_. It'd be nice to improve this in
the future, but will need a bit of threading-the-needle to convert the
pydantic settings classes to TypedDict and injecting those hints
reliably in all the right places.

---

The approach for the 'base' SharedConfig is as follows:

- All configuration should be declared here so that we can have one
  place to see all available settings.
- Default values should only be provided if they are: safe to be public,
  and suitable for a real production deployment. The latter is to guard
  against said value accidentally not being overridden. If we forget to
  override it for prod, we should fail as safely as possible.

I've added a couple of tests to help ensure that the base config and the
derived environment-specific configs stay in sync:

- types on the same setting must match
- environments cannot declare brand-new configurations that aren't on
  the shared base class.