FSPT-298: integration test to cover managed session commit
Checks that a data handler that doesn't commit itself is committed to
the database at the end of a request by integration testing the create
grant handler.

This surfaced that if an integrity check has already failed (and been
flushed to the database) the commit will fail with a rollback pending
error - we now handle that error and appropriately rollback but we want
to look into if theres a way of anticipating the rollback required.
Will follow up with a separate ticket/ commit.