Upgrade the DB in a separate process
When running migrations, alembic sources the `env.py` file, which can
have a bunch of side effects (like wiping logging configuration). It
really expects to be run from the CLI, which is an isolated (python)
environment, so it doesn't think about side effects.

Since we're running it programmatically, we have to think a bit about
side effects for global configuration if we're running it in the same
process. Let's avoid having to think about that by isolating it in a
separate python process.

We specifically set up the multiprocessing method as `fork`, which will
effectively create a clone of the Python process at the moment it's
called, so all of the context/configuration/etc will be the same (which
means things like the testcontainers DB URI will be available in the
child process).