Enable transactional test isolation
We want all of our integration tests to be isolated from each other, so
that any changes that happen in one cannot leak into others. We do this
by setting up a DB transaction before each test starts, and then rolling
that back at the end of the test. Any commits that happen within the
test itself are instead converted to SAVEPOINTS, which don't actually
commit to the DB.