FSPT-332: Move error summary to grant_base template
As we will have more forms in the Grant management area, it makes sense
to move the error summary to the manage_grant_base.html template so that
we don't have to keep adding it to each individual template and can
ensure consistent behaviour with all forms.

This commit also does some minor formatting tidying for all templates.