FSPT-332: Override main content block for backlinks
In order for backlinks to appear in the desired place on the page
when used in conjunction with the left-nav layout we need to override
the `{% block main %}` in the main govuk_frontend_jinja template to
accommodate our left nav layout.

This still allows backlinks to appear before the <main> HTML element
on the page, and tabbing to 'Skip to main content' works correctly.

We may want to revisit this implementation.