FSPT-298: FIX: make sure endpoints correctly commit integration tests
When writing integration tests we isolate each test using a function
scoped database transaction which is rolled back at the end of it.
Because of the save points this creates a `flush` to the database can be
treated as if its been committed between app requests, even though this
wouldn't work with the real app.

This fix proposes making sure that after each request to the app any
uncommitted changes are removed from the database session, this forces
the code under test to make sure its committing and will fail if
anything is missing (like the auto commit lifecycle decorator).