Precompile Jinja templates when running tests
We're trying to keep integration tests running quickly, and in order to
do that we've added a timeout as an autouse fixture. If an integration
test takes over x ms, we'll mark it as failed. We've had to bump this up
a few times since we first added it.

From profiling the tests, we found that the first time the test hits an
endpoint and has to render templates, Jinja needs to do a bunch of work
reading the files and compiling them. This only happens the first time,
so the first test run is artificially slower than it is on subsequent
runs, which has caused the first run to breach our timeouts.

By precompiling all of the templates as part of the app fixture setup,
we can take this compilation time out of the individual test run time,
which helps us stay within our timeout budget.

Without precompilation, we saw roughly these kind of times for the
`test_create_grant_handler` test:

0.21s call     tests/integration/test_handlers.py::test_create_grant_handler[1-2]
0.04s call     tests/integration/test_handlers.py::test_create_grant_handler[2-2]

After adding this precompilation step, the difference is drastically
reduced (but not a complete baseline so there may still be some cold
code paths to investigate later if we ever really want to):

0.08s call     tests/integration/test_handlers.py::test_create_grant_handler[1-2]
0.04s call     tests/integration/test_handlers.py::test_create_grant_handler[2-2]