BAU: tweak vite build settings
Tweaks the config to:
- stop the docker compose logs being hijacked by vite
- have the vite styles ready to go when booting up the dev server for
  the first time
- use the flask interface rather than dipping into node

The intention was to have the bundler split out the CSS instead of
having it load the SASS bundled in a dynamic JavaScript import. The
dynamic JavaScript import isn't cached by the browser and is what's
causing the flash of unstyled content. Splitting it out into a separate
CSS file would allow us to continue to get watch/ live reload but avoid
the FOUC. I wasn't able to appease it this time round.