FSPT-311: Remove flask-vite
Removes exsiting implementation of flask-vite. Adding similar functionality on
top/ replacing was getting confusing to keep track of. The app isn't in
a functional state as of this clearing out of files.

While flask vite took care of some of the install/ build/ asset linking
hooks for us it also traded off quite a lot of complexity where it
doesn't feel like we need it. It looks like it was designed for building
a Single Page Application with Flask as a backend - and therefore
assumes you'll have a single HTML entry point and import everything you
need into that. That model feels like a workaround in a static backend
driven web app and is quite hard to understand if you're not familar
with the tools. Using vite directly should allow us to directly point it
add SCSS/ JS that needs to be transpiled and manage those assets
directly.