FSPT-311: adds vite dependency directly to repository
Include vite/ npm dependencies at the package level, this make it feel
more like a build process for the app and makes it clear that these are
things that are _required_ to build and run the app.

For parity with the existing build this is what will take care of
bringing in the latest version of the design system and keeping us up to
date with that.

Ideally we should be able to keep these dependencies up date using
rennovate.

This commit sets up the dev server and directly points templates at
it. This should enable a good developer experience (HMR, page reloads,
etc.) but doesn't yet support the static "production" build.

By pointing directly at the SCSS thats being built on the dev server we
avoid the "Flask of unstyled content" which was being caused by loading
JavaScript that would then add styles to the page (with the previous SPA
setup).