FSPT-311: hook vite build into docker compose
Update the different build contexts to use the method of calling vite
directly. It should only be in a local environment that we're pointing
at and running the dev server but it looks like we currently do that for
pull previews - we'll need to decide how we want this to behave going
forward.