fixup: don't clobber the images node modules folder when mounting
I have no idea why this wasn't happening to me before - maybe it was
using a cached volume in a way that I don't quite understand.

We don't want to rely on the host machines node modules to
a) allow the host machine to only worry about python
b) avoid issues with x-architecture installs

Instruct docker compose to mount the app folder but ignore node modules,
instead favouring the node modules that was installed with the image.