Generate and claim magic links
Generate a MagicLink record when a user submits their email address on
the 'request a link to sign in' page. Inject the magic link information
into the Notify email sent.

Adds a stub page for claiming the magic link. A future commit will add
JS to auto-submit the form so we wouldn't expect many users to see
anything here - they should get automatically redirected to the page
they were trying to access before signing in.