BAU: consistent back link behaviour
Updates any forms to strip text input and start adding back links to
pages.

We'll need to update the page layout `main-wrapper` behaviour to
independently wrap the nav and the main content so that we don't have to
do back link spacing workarounds - this will be a follow up PR.

We should do this consistently using base forms when we have enough use
cases but this will do for now.