Autouse DB fixture; remove redundant fixtures
The `app` fixture was relying on `setup_db_container`, which spins up
postgres but does not run migrations. Anything that needs the db to
exist should be relying on the `db` fixture.

This patch changes app to rely on the `db` fixture, which gives it a
migrated db.

I've also set the `db_session` fixture as an autouse one. This means
every single test in the `integrations` folder will always have a
transactionally-isolated session. It means a slight bit of
waste/overhead on any 'integration' tests that don't need the db, but
the confidence it gives us that future developers will always be writing
tests with in a safe space feels worth this for now.