FSPT-278: Make e2e test only success path & easier to read
To make the tests more readable, moved the `expect` statements to be
part of the helper methods that encapsulate page actions.

For static locators that will rarely change, let's also make them
properties of the class rather than needing a method to get them,
so we can keep methods for dynamic locators or for repeatable actions
on the page.

As part of this, also remove the expects around error messages - these
are and should be covered as part of unit/integration tests, and e2e
tests can therefore focus on the success path and the actions and methods
follow user actions more closely.

When running e2e tests locally, Flask Debug Toolbar will be on by default
which can hide some of the elements on the page that Playwright is trying
to locate and interact with. We also ran into this issue on old
post-award and so setting the viewport size to be larger by default
gets around this issue.