FSPT-278: Make errors mixin and static locators properties
Error summaries are going to be used a lot so it makes sense for these
to be abstracted and reusable for assertions. Turning it into a mixin
with methods that accept a string for the error title or subtitle
should also make the tests themselves readable and avoid having to
flit back and forth between the test and Page definitions.

For static locators that will rarely change, let's also make them
properties of the class rather than needing a method to get them,
so we can keep methods for dynamic locators or for repeatable actions
on the page.