Auto-claim magic link via JS
We don't need users to click a button manually when using their magic
link. The reason we add a button is to prevent email clients, which may
preload the magic link URL for various reasons, from expiring the magic
link. Email clients won't run JS on the page, so this shouldn't
undermine the reason for us having this interstitial page.