Add Flask-Session using SQLAlchemy
Move from using Flask's native session implementation, which stores all
data in a session cookie, to using server-side cookies that only put an
ID in the session cookie. All other data is stored in a `sessions` table
in our postgres DB.

Flask-Session does not currently support Flask-SQLAlchemy-Lite, the
slimmer version of Flask-SQLAlchemy we've decided to use in this app, so
I've had to patch Flask-Session to add this support.

If we keep this approach, we need to get this patch upstream into
Flask-Session and retire the fork ASAP.