Add Flask-Login
Implement login-with-magic-link functionality using Flask-Login. We hook
up our User model to the extension and add a loader that will allow
Flask-Login to resolve a user's session/id to a real User instance. This
gets made available to the rest of the app (and templates) as
`current_user`.

Flask-Login handles authentication only - it does not care about
authorization. We will need to handle this separately when we get to
that point.