Remove sign in; add sign out
Remove the 'sign in' link; we will redirect users to the sign in page if
they try to access a protected resource.

However, users should be able to sign out at will, so we'll add a sign
out link if they're logged in.

We'll just do this as a GET for now. This does mean that users could get
'tricked' into logging out, but the damage here is low (and GOV.UK
Notify and GOV.UK Pay do the same; a dangerous precedent indeed).