FSPT-357: hide the interstitial form if js is enabled
This prevents a very quick flash of green when loading and immediately
clicking the interstitial page following a magic link.

We'll need to decide if we want this interstitial page to sit in the
same service nav structure as the rest of the app (as it currently does)
but that feels like a decision to make when there are more contexts to
consider.

`display: none` will only be applied if gov uk frontend jinja was able
to set a `js-enabled` class on the whole page.
Reference:
https://github.com/LandRegistry/govuk-frontend-jinja/blob/63864e76095323d72357edd68463aaf4597e4763/govuk_frontend_jinja/templates/template.html#L29