FSPT-278: Fix viewport size issue
When running e2e tests locally, Flask Debug Toolbar will be on by default
which can hide some of the elements on the page that Playwright is trying
to locate and interact with.

We also ran into this issue on old post-award and so setting the viewport
size to be larger by default gets around this issue.