FSPT-278: Allow E2E tests to run against dev env
Passing the `--e2e-env dev` flag allows you to run the e2e tests against
the dev environment.

This also now allows the e2e tests to run against dev as part of the
CI/CD pipeline or in their own workflow.

The CI/CD e2e tests are currently commented out as until we have a nice
way of checking that a deploy has successfully completed and can run the
e2e tests this behaviour will be very flakey.