BAU: constrain test and CI postgres versions
Our workflows/ CI/ test environment should aim to reflect the deployed
environments as much as possible to give us confidence that automated
checks are relevant to what will then be run.

This attempts to tell renovate to give us pinned patch upgrades but keep
the major/ minor version of postgres the same as our deployed version,
this should then be bumped when minor/ major upgrades are made (this
needs to be sense checked against our DB upgrade policy, I believe its
constrained to minor versions but if there's an automatic upgrade policy
the rules here may need to change).

I'm not confident in the renovate syntax proposed here that may take
some trial and error.