Update e2e tests for magic link journey
This pulls in a lot of frameworking from our existing playwright e2e
test suite in pre- and post-award, specifically the idea of:

- EndToEndTestSecrets
- Reading from AWS for deployed envs
- Scraping magic links from GOV.UK Notify
- A `user_auth` fixture that will start most e2e tests already logged-in

I've separated the magic link journey out into a small separate e2e test
because this has a bunch of redirect logic and auto-executing JS that is
otherwise hard to test as a cohesive journey.

With this covered, all of the other e2e tests should be able to ignore
this and start from a logged-in place. This is facilitated by the
`user_auth` fixture, which when used in an e2e test will step through
the magic link journey for us. An email address will be automatically
generated based on the name of the test and a small random fragment. The
domain for the email can be overridden with the `user_domain` mark.