FSPT-377: Make AppRunner deploy manual
We want visibility and control over the AppRunner deployments rather
than relying on automatic deployments when a new image with the `latest`
tag is pushed to ECR.

We're not sure if we'll continue to use AppRunner so for the time being
we can use the AWS AppRunner deploy action and specify deploying
the image we've just built in the pipeline. This allows us to then
successfully run the e2e tests after a successful deploy or catch
any issues with deployments (which before now could fail silently).

On merge we're saving the image as `:main` and tagging with the github
sha and `latest`, but this means if you try to build and push an image
from a branch eg `bau/<some_branch_name>` the build fails because of the
slash in the branch name.

This changes to just save the images with the github sha directly and
only tag as latest if it's a merge to main.