FSPT-377: Make AppRunner deploy manual
As part of FSPT-377 we want visibility and control over the AppRunner
deployments rather than relying on automatic deployments when a new
image with the `latest` tag is pushed to ECR.

Now let's use the AWS AppRunner deploy action and specify deploying
the image we've just built in the pipeline. This should also then allow
us to successfully run the e2e tests after a successful deploy or catch
any issues with deployments (which before now could fail silently).