FSPT-379: authorise once per test across distributed workers
This proposes setting up a file lock in the user auth fixutre for
end to end tests. This allows any number of workers to sping up and
re-use a session which is correctly authorised once.

Currently all end to end tests that pull in `user_auth` will create a
new random user in the system and take the time to go through the magic
link process individually.

This change means even if running pytest with `-n auto` the authorised
session can be used.

Note that this doesn't guard against the scenario when you're not
running in parallel mode as the fixture is function scoped which would
result in all tests authorising individually - this behaviour doesn't
feel quite right so we should interrogate if we can reproduce this
behaviour without manipulating the individual `page` fixture.