BAU: cleanup docker compose startup logs
We're unlikely to be debugging core modules which means it should be OK
to use the Python frozen modules optimisation - we're doing that as is.
If we're OK to continue doing that shut off the warning (alternatively
we can tell Python to not freeze modules at a slight startup speed
penalty).

Also remove logs that don't feel like they add much.