Tweak logging for Sentry
Sentry does not integrate with our app logging config, instead working
at a slightly higher level - so the logs printed to stdout by our config
are not the logs shipped to Sentry.

This isn't great, but for now Sentry logging is probably still a
slightly better win than nothing.

This tweaks the logs in two ways:
- disable the werkzeug logger in Sentry explicitly; we disable this in
  our own app logging config. Annoying to duplicate this, but fine for
  now.
- Tweak the log messages for request received/response sent logs so that
  they stack on top of each other nicely for reading/scanning.