FSPT-377: Update image tagging
On merge we're saving the image as `:main` and tagging with the github
sha and `latest`, but this means if you try to build and push an image
from a branch eg `bau/<some_branch_name>` the build fails because of the
slash in the branch name.

This changes to just save the images with the github sha directly and
only tag as latest if it's a merge to main.