FSPT-377: Push to GHCR and then to ECR in pipeline
Now that we have multiple deployed environments (currently dev and test)
we run into the issue that each AWS account has its own ECR. We only
want to build our images once and deploy this same image to each
environment in turn, which becomes tricky when we push the built image
to the ECR of the lowest environment first.

We could allows each account to have access to the ECR of the account/
environment beneath it, but this would require updates to the terraform
setup and injecting unique account IDs conditionally (as you'd only want
test to be able to access dev, and prod to access test etc.), and feels
like it could run into some rough edges.

We discussed having a separate AWS CI account which would host all the
built images and then our pipeline would pull the image to that
environment's ECR and deploy it. While we wait to have a separate AWS CI
account set up, we agreed to use GitHub Container Registry as an
interim solution, as it's what we already use in our existing services.