FSPT-379: e2e lookup magic links by id
Adds an ephemeral reference to the notification sent during the magic
link auth to the page using a `[data-x]` attribute.

Remove the logic where we generate random unique email addresses in then
be able to look them up in Notify. Instead directly tie the notification
that was sent in that session to the page checking the email.

This will now let us use one email address that actually sends emails,
while we're running end to end tests against actual environments this
allows us to use an API key that sens email allowing pre-production
users to not log into the Notify admin console.

The method should also work without a problem for test Notify keys that
pretend to send emails.

We're making the choice to only include the `[data-x]` attribute if we're
not in a production Flask environment but that can be removed if
neccessary - there should be nothing sensitive about the emails
reference.

A follow up PR will then switch the fixture providing `user_auth` for
end to end tests to share a single session rather than uniquely
authenticating every test.