Fix fixture resolution
`app` was, probably unexpectedly, depending on `db`, while `db` was
depending on `app`. I'm not sure how pytest ever resolved this, but
after adding the `test_app` file that checks basic auth, which uses
another app fixture called `app_with_basic_auth`, pytest started
throwing fixtures resolution errors because of this cyclical dependency.

All `app` requires is for the DB config to be pushed into the
environment; the `db` fixture requires the `app` fixture in order to get
an app context to read config. So the flow should be, and now is:

setup_db_container -> app -> db