Patch environment proximal to app creation
We've been patching the environment to set some variables that contain
eg our DB config, and then consuming that in other fixtures. Depending
on the order the fixtures are then used by pytest, the environment could
end up in an inconsistent state (because we have some tests that patch
the environment with a valid DB and others that patch it with an invalid
DB).

This patch moves all patching of the environment into the same
fixture/function/location where we build the app itself, which should
make it much more consistent and reliable.

This problem only manifested when running the tests in a random order,
either by installing and using pytest-random-order, or by running the
tests in multiprocess mode (as we do in github CI). This problem also
only occasionally manifests, leading to flaky tests that don't often
show up locally.