FSPT-341: adds prettier formatter
This commit adds the prettier formatter to the node package manager
dependencies. This configuration for fomatting jinja templates is the
only one we've found so far that feels like it does a consistent job.

As we're not yet convinced this commit doesn't include it in pre-commit
or CI breaking checks but just makes it available to developers to use
if they're working on templates and would like opinions.