WIP: FSPT-379: moves parallel fixture sharing to session scope
The `browser` fixutre provided by playwright is set at the session scope
which means we should be use it to set these flows up only once at the
start.

This attempt continues to be syncronous - I think its only worth looking
into this when there are more tests and I'm not just checking things
against a `--count=10` which might behave differently anyway.