BAU: Fix `updated_at` behaviour
We're using `server_onupdate` for our `updated_at` column in the base
model to catch timestamps when rows in our database are updated.

However this doesn't seem to be firing and all our `updated_at` columns
in tables that inherit from the base model are still set at the same
time as `created_at`, even after updates.

The reason for this seems to be that `server_onupdate` doesn't
automatically trigger when row is updated, and instead needs explicit
instruction to update that column during an UPDATE operation eg. with
a trigger.

Switching to just using `onupdate` allows SQLAlchemy to handle this
behaviour when an entity is updated through the ORM.