Set SERVER_NAME to support cloudfront
We're putting cloudfront in front of apprunner now, and hooking that up
to a custom funding.*.communities domain. Cloudfront will not forward
that hostname to us, instead sending us CloudFront's own domain. So our
app needs to know what domain is actually public facing so that it can
build the correct URLs for eg magic links.