Update the TimeFreezer to support moving times
Update the function injected into the test DB so that it reads the
'frozen time' from a setting; this lets us then just change the value of
the setting to move time in the DB, rather than recreating the function
entirely each time.

Drops a test for updated_at working when time is not frozen (because we
run each test in a transaction, and transactions only have a single
timestamp - the start of the transactions).

Update column names from a rebase.