FSPT-388: Implement basic SSO
Add basic SSO login to the service. This doesn't yet add any role
information to the DB, which will be done as part of a subsequent commit.

Because we're using the same app in Azure as the existing services,
the redirect URI needs to match the one already registered in the
Azure portal which is why the route needs to stay as `/sso/get-token`.