FSPT-388: Add SSO stub server for SSO e2e tests
For integration and unit tests we can mock out SSO but in order to test
the full SSO sign-in journey with e2e tests we need to be able to
consistently use or replicate the SSO sign-in journey.

For local e2e tests, we are doing this by implementing a stub service
which implements the same SSO interface and flow and "returns" an
authenticated user.

This is trickier on deployed environments where we are looking into
having a test user provisioned by IT that we can use to do 'real' logins
during some browser-based flow.