FSPT-388: Add tests for SSO sign-in journey
Adds a small integration test for the SSO sign-in page and an e2e test
for the full SSO sign-in journey. This test uses the stub server when
running the tests locally and will currently be skipped in deployed
environments until we do the work to consistently login properly to
Microsoft with some kind of service account and input/bypass MFA on
those environments.