BAU: Allow @test.communities email domain
Now we've implemented SSO, people will be signing in with their
@test.communities.gov.uk accounts through SSO. As we're only using the
test key currently, this PR updates our form validation and route
wrapper to allow logins with both @communities and @test.communities
emails.

For consistency this has also now been applied to the magic link sign-in
page, which previously validated to only allow @communities emails as
well. It's likely we'll drop this validation in future once the platform
is more built out, as this sign-in route will be the one that's open to
and used by most external users, but for now internal users can
continue to sign in with both methods.