BAU: Add dummy default Azure AD values to PullPreview config
The PullPreview needs dummy default values for the Azure AD config
variables which are required for SSO, in the same way that the LocalConfig
required defaults.

PullPreview won't currently work with SSO as this needs to be integrated
into the dockercompose for and start-up for pullpreviews, but this
will at least allow them to build and start up correctly and magic
link sign-in to be used until SSO is properly integrated.