Refactor: move config to module
Intellisense/goto file isn't quite as nice a developer experience when
they're inside config/__init__.py as opposed to being in a file just
called config.py, so let's move them. Given the config is just a few
classes and not a sprawling set of distinct things, we probably don't
need it to be a full package.