Refactor: move healthcheck to module
Intellisense/goto file isn't quite as nice a developer experience when
they're inside healthcheck/__init__.py as opposed to being in a file just
called healthcheck.py, so let's move them. Given the healthcheck is just
one blueprint/route and not a sprawling set of distinct things, we
probably don't need it to be a full package.