Add meta-test for error summary rendering
We 'automate' error summary rendering by looking for a 'form' argument
passed into the template. If this exists, then we check 'form.errors' -
and if there are any, we render the error summary in a standard place on
the template.

This patch adds an assertion onto the flask test client that checks the
rendered templates. If it receives a `form` argument, we validate that
it's actually a form instance, rather than eg a DB object.