FSPT-410: Add UserRole and Organisation tables
Creates UserRole table for storing a User's roles at all levels. By
having nullable columns for Org and Grant Foreign Key references, we can
infer the permission level for that user, eg. if they have the `admin`
role but org and grant are null, then they are a global/platform admin.
If they have the `admin` role and an org id, then they are an
organisation-level admin who would be able to eg. add other users to the
organisation. This also means we just have one table to check for
roles and permissions rather than splitting it across multiple join
tables.

Adds some basic check constraints for now for obvious cases.

In order to make this work for now, this commit also adds a stub
Organisation table, which will certainly be extended in the future but
means we can have the UserRole table in place.