FSPT-410: Make nulls not distinct for UserRole unique constraint
By default, NULL values are treated as distinct entries by Postgres.
Specifying NULLS NOT DISTINCT on unique indexes / constraints will cause
NULL to be treated as not distinct, which is what we want on the UserRole
table where the NULL values are important information and should not be
distinct (otherwise a platform admin user could have a new row added to
the table each time they log in).

This is a small change to make NULL values not distinct for the UserRole
table so this unique constraint can be enforced.