Use composite PK on CollectionSchema
https://mhclgdigital.atlassian.net/browse/FSPT-437

The CollectionSchema table will (at some future time) support
versioning. When a new version of the schema is created, we want a
stable identifier that will let us track versions over time. This will
be the `id` column. The `version` column will get incremented each time.

At the moment the `id` column is the sole PK, which would require later
versions to have a new ID, and make it hard to track versions over time.

This patch switches us to using a composite PK (id, version) and
cascades that appropriately to the related `sections` table's FK.

Rather than writing a migration that applies this cleanly, I think it's
OK for now to manually wipe out our databases and then re-build them
from scratch.