Rename 'collections' to 'schemas' in the scaffolding
We're about to let developers start collecting data on a collection
schema, and I think that calling the schemas 'collections' is a bit
backwards. This patch renames those to 'schemas', and then the actual
instance of a user submitting data can be called a 'collection'.

Note: these words don't matter too much at the moment, just for while
we're scaffolding.