FSPT-437 Scaffold the tasklist page for a collection
This patch adds the bones of a tasklist view for an instantiated
collection: list all of the sections and forms, although you can't yet
fill them in.

The statuses are hard-coded and will handled as part of a future ticket.

Adds a 'preview collection' button to the 'manage (collection) schema'
developers page.